model.builder.upload
:)

+++ Hackathon_application/HA.Domain/Enitites/PriceMultipliers.cs
@@ -10,7 +10,7 @@ namespace HA.Domain.Enitites
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public float Multiplier { get; set; }
+        public double Multiplier { get; set; }
         public Guid ThresholdFirstId { get; set; } // PM 10
         public Guid ThresholdSecondId { get; set; } // PM 2.5
     }
